% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LongitudinalData.R
\docType{class}
\name{LongitudinalData-class}
\alias{LongitudinalData-class}
\title{class LongitudinalData}
\description{
A class for LongitudinalData.
}
\section{Slots}{

\describe{
\item{\code{id}}{a character representing subject ID}

\item{\code{visit}}{an integer corresponding to the visit number (0, 1, 2)}

\item{\code{room}}{a character representing the room name}

\item{\code{value}}{a numeric representing the pollution measured in micrograms per cubic meter}

\item{\code{timepoint}}{an integer representing the time the measurement was taken}

\item{\code{selectiontags}}{a character}
}}

