% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LongitudinalData.R
\name{make_LD}
\alias{make_LD}
\title{LongitudinalData Constructor}
\usage{
make_LD(DF)
}
\arguments{
\item{DF}{a data frame}
}
\value{
LD a LongitudinalData object
}
\description{
This function constructs a new LongitudinalData object given a data frame.
}
