% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LongitudinalData.R
\docType{methods}
\name{print,LongitudinalData-method}
\alias{print,LongitudinalData-method}
\title{print}
\usage{
\S4method{print}{LongitudinalData}(x)
}
\arguments{
\item{x}{a LongitudinalData object}
}
\value{
None
}
\description{
This function prints a description of a LongitudinalData object.
}
