% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LongitudinalData.R
\name{room}
\alias{room}
\title{room}
\usage{
room(LD, room_char)
}
\arguments{
\item{LD}{a LongitudinalData object}

\item{room_char}{a character representing the room name}
}
\value{
room a LongitudinalData object
}
\description{
A generic function for extracting room-specific information.
}
