% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LongitudinalData.R
\name{subject}
\alias{subject}
\title{subject}
\usage{
subject(LD, subject_char)
}
\arguments{
\item{LD}{a LongitudinalData object}

\item{subject_char}{a character representing subject ID}
}
\value{
subject a LongitudinalData object
}
\description{
A generic function for extracting subject-specific information.
}
