% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LongitudinalData.R
\docType{methods}
\name{summary,LongitudinalData-method}
\alias{summary,LongitudinalData-method}
\title{summary}
\usage{
\S4method{summary}{LongitudinalData}(object)
}
\arguments{
\item{object}{a LongitudinalData object}
}
\value{
list containing Subject ID string and summary object
}
\description{
This function returns mean summary statistics for a LongitudinalData object.
}
