% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LongitudinalData.R
\name{visit}
\alias{visit}
\title{visit}
\usage{
visit(LD, visit_int)
}
\arguments{
\item{LD}{a LongitudinalData object}

\item{visit_int}{an integer corresponding to the visit number (0, 1, 2)}
}
\value{
visit a LongitudinalData object
}
\description{
A generic function for extracting visit-specific information.
}
